Refactor ir tests cases to unfold test cases
