+++ test/wat_compiler/resources/function_expressions.testcase.ts
