+++ test/wat_compiler/resources/unfold.testcase.ts
@@ -1,19 +1,18 @@
-import { type Token } from '../../../src/common/token';
+/**
+ * Test cases for testing unfolding of intermediate representations
+ */
 import {
   OperationTree,
   PureUnfoldedTokenExpression,
   UnfoldedTokenExpression,
-  type IntermediateRepresentation,
+  type Unfoldable,
 } from '../../../src/wat_compiler/ir_types';
-import { type Tree } from '../../../src/wat_compiler/tree_types';
 import { getSampleToken as t } from './resolved_tokens';
 
 export interface FunctionBodyTestCaseData {
   str: string;
-  parseTree: Tree<Token>;
-  ir: IntermediateRepresentation;
+  ir: Unfoldable;
   unfolded_ir: PureUnfoldedTokenExpression;
-  minimal_binary: Uint8Array;
 }
 
 const simple_addition_sexpr: FunctionBodyTestCaseData = {
@@ -23,11 +22,6 @@ const simple_addition_sexpr: FunctionBodyTestCaseData = {
         (f64.const 1.5)
         )
     `,
-  parseTree: [
-    t('f64.add'),
-    [t('f64.const'), t('1')],
-    [t('f64.const'), t('1.5')],
-  ],
   ir: new OperationTree(t('f64.add'), [
     new UnfoldedTokenExpression([t('f64.const'), t('1')]),
     new UnfoldedTokenExpression([t('f64.const'), t('1.5')]),
@@ -35,9 +29,6 @@ const simple_addition_sexpr: FunctionBodyTestCaseData = {
   unfolded_ir: new PureUnfoldedTokenExpression(
     ['f64.const', '1', 'f64.const', '1.5', 'f64.add'].map(t),
   ),
-  minimal_binary: new Uint8Array([
-    0x44, 0, 0, 0, 0, 0, 0, 0xf0, 0x3f, 0x44, 0, 0, 0, 0, 0, 0, 0xf8, 0x3f, 0xa0,
-  ]),
 };
 
 const simple_addition_stack: FunctionBodyTestCaseData = {
@@ -46,16 +37,12 @@ const simple_addition_stack: FunctionBodyTestCaseData = {
         f64.const 1.5
         f64.add)
     `,
-  parseTree: ['f64.const', '1', 'f64.const', '1.5', 'f64.add'].map(t),
   ir: new UnfoldedTokenExpression(
     ['f64.const', '1', 'f64.const', '1.5', 'f64.add'].map(t),
   ),
   unfolded_ir: new PureUnfoldedTokenExpression(
     ['f64.const', '1', 'f64.const', '1.5', 'f64.add'].map(t),
   ),
-  minimal_binary: new Uint8Array([
-    0x44, 0, 0, 0, 0, 0, 0, 0xf0, 0x3f, 0x44, 0, 0, 0, 0, 0, 0, 0xf8, 0x3f, 0xa0,
-  ]),
 };
 
 const nested_addition_stack: FunctionBodyTestCaseData = {
@@ -68,19 +55,6 @@ const nested_addition_stack: FunctionBodyTestCaseData = {
       f64.add
       f64.add
       `,
-  parseTree: [
-    'f64.const',
-    '1',
-    'f64.const',
-    '1',
-    'f64.add',
-    'f64.const',
-    '1',
-    'f64.const',
-    '1',
-    'f64.add',
-    'f64.add',
-  ].map(t),
   ir: new UnfoldedTokenExpression(
     [
       'f64.const',
@@ -111,47 +85,6 @@ const nested_addition_stack: FunctionBodyTestCaseData = {
       'f64.add',
     ].map(t),
   ),
-  minimal_binary: new Uint8Array([
-    0x44,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0xf0,
-    0x3f,
-    0x44,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0xf0,
-    0x3f,
-    0xa0,
-    0x44,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0xf0,
-    0x3f,
-    0x44,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0xf0,
-    0x3f,
-    0xa0,
-    0xa0,
-  ]),
 };
 
 const nested_addition_sexpr: FunctionBodyTestCaseData = {
@@ -167,11 +100,6 @@ const nested_addition_sexpr: FunctionBodyTestCaseData = {
       )
     )
       `,
-  parseTree: [
-    t('f64.add'),
-    [t('f64.add'), [t('f64.const'), t('1')], [t('f64.const'), t('1')]],
-    [t('f64.add'), [t('f64.const'), t('1')], [t('f64.const'), t('1')]],
-  ],
   ir: new OperationTree(t('f64.add'), [
     new OperationTree(t('f64.add'), [
       new UnfoldedTokenExpression([t('f64.const'), t('1')]),
@@ -197,47 +125,6 @@ const nested_addition_sexpr: FunctionBodyTestCaseData = {
       'f64.add',
     ].map(t),
   ),
-  minimal_binary: new Uint8Array([
-    0x44,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0xf0,
-    0x3f,
-    0x44,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0xf0,
-    0x3f,
-    0xa0,
-    0x44,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0xf0,
-    0x3f,
-    0x44,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0xf0,
-    0x3f,
-    0xa0,
-    0xa0,
-  ]),
 };
 
 export const validTestCases = [
@@ -245,5 +132,4 @@ export const validTestCases = [
   simple_addition_stack,
   nested_addition_stack,
   nested_addition_sexpr,
-  // simple_addition_sexpr_without_argument_bracket_fails,
 ];
