+++ test/wat_compiler/unit.test.ts
@@ -1,11 +1,13 @@
 import { type Token } from '../../src/common/token';
+import { type ModuleExpression } from '../../src/wat_compiler/ir_types';
+import { type ParseTree } from '../../src/wat_compiler/tree_types';
+
 import { BinaryWriter } from '../../src/wat_compiler/binary_writer';
 import { getIR } from '../../src/wat_compiler/ir';
-import { type ModuleExpression } from '../../src/wat_compiler/ir_types';
 import { tokenize } from '../../src/wat_compiler/lexer';
 import { getParseTree } from '../../src/wat_compiler/parser';
-import { type ParseTree } from '../../src/wat_compiler/tree_types';
-import { positiveFunctionTestCases } from './resources/function_expressions.testcases';
+
+import { positiveFunctionTestCases } from './resources/function_expressions.testcase';
 
 describe.each(positiveFunctionTestCases)(
   'unit tests for function expressions',
@@ -13,12 +15,13 @@ describe.each(positiveFunctionTestCases)(
     let tokens: Token[];
     let parseTree: ParseTree;
     let IR: ModuleExpression;
-    let binary: () => Uint8Array;
+    let binary: Uint8Array;
     beforeEach(() => {
       tokens = tokenize(testCase);
       parseTree = getParseTree(tokens);
       IR = getIR(parseTree) as ModuleExpression;
-      binary = new BinaryWriter(IR).encode;
+      binary = new BinaryWriter(IR)
+        .encode();
     });
     test('test lexer', () => {
       expect(tokens)
